requirements.txt
Flask
seaborn
matplotlib
PyYAML